Fix Junction Structure UI: remove Advanced expander, fix max_earth_cover validation
- Change max_earth_cover_ft from float to int (values 10 or 20 are whole numbers)
- Move all Junction Structure inputs to primary (remove Advanced box)
- Fix InputField validation to properly cast choice values to dtype after validation
  Choices are always strings in schema; validation converts value to string for comparison,
  then casts the string choice to the proper dtype (e.g., int("10") = 10)

Fixes: max_earth_cover_ft: '10.0' not in ['10', '20'] validation error

+++ vistadetail/engine/templates/junction_structure.py
@@ -37,7 +37,7 @@ class JunctionStructureTemplate(BaseTemplate):
                        min=5.5, max=12.0, default=5.5,
                        group="Box Geometry",
                        hint="Inside height from floor to top slab soffit — D91B range 5'-6\" to 12'"),
-            InputField("max_earth_cover_ft", float, label="Max Earth Cover (ft)",
+            InputField("max_earth_cover_ft", int, label="Max Earth Cover (ft)",
                        choices=["10", "20"], default="10",
                        group="Loading",
                        hint="Maximum earth cover over top slab — governs D91B table row (10 or 20 ft)"),
